--- platform/viewer/src/containers/StudyListContainer/index.js
@@ -1,1 +1,0 @@
-export { default } from './StudyListContainer';
