--- platform/viewer/src/containers/ViewerContainer/ViewerContainer.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-const ViewerContainer = () => {
-  return (
-    <div>
-      <div>ViewerContainer</div>
-      <Link to="/">Go to StudyList Route</Link>
-    </div>
-  );
-};
-
-export default ViewerContainer;
