--- platform/viewer/src/containers/ViewerContainer/index.js
@@ -1,1 +1,0 @@
-export { default } from './ViewerContainer';
